requirements.txt
streamlit
pypdf
faiss-cpu
sentence-transformers
torch
transformers
numpy
